requirements.txt
# Direct dependencies are declared in pyproject.toml.
# This file is the pinned lock file (pip freeze) for reproducible installs.
#
# Install:  pip install -r requirements.txt
# Or:       pip install .              (uses pyproject.toml)
# Or:       pip install ".[dagster]"   (includes Dagster orchestration)
# Or:       pip install ".[dev]"       (includes pytest + sqlfluff)

adlfs==2025.8.0
agate==1.9.1
aiobotocore==2.24.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
arro3-core==0.6.3
async-timeout==5.0.1
attrs==25.4.0
azure-common==1.1.28
azure-core==1.35.1
azure-datalake-store==0.0.53
azure-identity==1.25.0
azure-mgmt-core==1.6.0
azure-mgmt-storage==23.1.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
babel==2.17.0
botocore==1.40.18
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
daff==1.4.2
dbt-adapters==1.17.2
dbt-common==1.32.0
dbt-core==1.10.13
dbt-duckdb==1.9.6
dbt-extractor==0.6.0
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
deltalake==1.1.4
Deprecated==1.2.18
diff_cover==9.7.1
dlt==1.17.1
duckdb==1.4.0
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
giturlparse==0.12.0
hexbytes==1.3.1
humanize==4.13.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
Logbook==1.8.2
loguru==0.7.3
makefun==1.16.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.7.0
networkx==3.5
numpy==2.3.3
ordered-set==4.1.0
orderly-set==5.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.1.0
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
polars==1.34.0
polars-runtime-32==1.34.0
portalocker==3.2.0
propcache==0.4.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requirements-parser==0.13.0
rich==14.1.0
rich-argparse==1.7.1
rpds-py==0.27.1
s3fs==2025.9.0
semver==3.0.4
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
sqlalchemy==2.0.48
sqlfluff==3.4.2
sqlglot==27.20.0
sqlparse==0.5.3
tblib==3.1.0
tenacity==9.1.2
text-unidecode==1.3
time-machine==2.19.0
tomlkit==0.13.3
tqdm==4.67.1
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
